requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cloudinary==1.20.0
dj-database-url==0.5.0
Django==3.0.6
django-debug-toolbar==2.2
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
